+++ CryonicEngine/RaylibDrawWrapper.h
@@ -2,7 +2,7 @@
 
 void DrawRectangleWrapper(float posX, float posY, float sizeX, float sizeY, float rotation, unsigned char colorR, unsigned char colorG, unsigned char colorB, unsigned char colorA);
 
-void DrawCircleWrapper(int posX, int posY, float radius, unsigned char colorR, unsigned char colorG, unsigned char colorB, unsigned char colorA);
+void DrawCircleWrapper(float posX, float posY, float radius, unsigned char colorR, unsigned char colorG, unsigned char colorB, unsigned char colorA);
 
 void DrawTextureWrapper();
